Config in Video Player
